refactor: move notifications infrastructure to shared components
