+++ frontend/src/components/notifications/components/NotificationManager.tsx
