+++ frontend/src/components/notifications/hooks/useNotificationManager.ts
