+++ frontend/src/features/auth/components/RegisterForm.tsx
@@ -8,7 +8,7 @@ import {
 } from "@/features/auth/components/RegisterFormSteps";
 import { StepIndicator } from "@/features/auth/components/StepIndicator";
 import { REGISTRATION_STEPS } from "@/features/auth/utils";
-import FloatingNotification from "@/features/notifications/components/FloatingNotification";
+import FloatingNotification from "@/components/notifications/components/FloatingNotification";
 import { useFeatureLoading, useMutationErrorHandler } from "@/hooks";
 import { useStore } from "@/store/store";
